Create shell_plus-get_modified_files
I ran into a need to identify hooks modified in a CloudBolt v7 (py2) to know which files to extract and port over to CloudBolt v9 (py3). 

I was not sure how to query CloudBolt for objects modified, so I took the route I did know and queried the underlying OS file modifications and compared those paths to CloudBolt hook paths.